+++ src/main/java/fr/honeygroup/bo/response/LeadResponse.java
@@ -8,14 +8,14 @@ import java.util.Map;
 @Data
 @Builder
 public class LeadResponse {
-    private Integer id;
+    private Long id;
     private LocalDateTime dateSoumission;
     private String statut;
     private String source;
-    private Integer userId;
+    private Long userId;
     private String userNomComplet; // Pratique pour l'affichage
-    private Integer poleId;       // Ajouté car important dans ton SQL
-    private Integer prestationId;
+    private Long poleId;       // Ajouté car important dans ton SQL
+    private Long prestationId;
     private String prestationTitre;
     private Map<String, String> details;
 }
