+++ src/main/java/fr/honeygroup/repository/DemandeLeadRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DemandeLeadRepository extends JpaRepository<DemandeLead, Integer> {
+public interface DemandeLeadRepository extends JpaRepository<DemandeLead, Long> {
 
     /**
      * Pour le dashboard commercial : voir les leads par état.
@@ -18,12 +18,12 @@ public interface DemandeLeadRepository extends JpaRepository<DemandeLead, Intege
     /**
      * Pour l'historique d'un client spécifique.
      */
-    List<DemandeLead> findByClientId(Integer clientId);
+    List<DemandeLead> findByClientId(Long clientId);
 
     /**
      * Pour filtrer les demandes par pôle (ex: toutes les demandes "Boxe").
      */
-    List<DemandeLead> findByPoleId(Integer poleId);
+    List<DemandeLead> findByPoleId(Long poleId);
 
     /**
      * Bonus : Trouver les demandes les plus récentes en premier.
