+++ src/main/java/fr/honeygroup/repository/DetailsSpecifiquesRepository.java
@@ -7,13 +7,13 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DetailsSpecifiquesRepository extends JpaRepository<DetailsSpecifiques, Integer> {
+public interface DetailsSpecifiquesRepository extends JpaRepository<DetailsSpecifiques, Long> {
 
     /**
      * Récupère tous les détails liés à une demande spécifique.
      * Utile si tu veux recharger uniquement les détails sans recharger tout le Lead.
      */
-    List<DetailsSpecifiques> findByDemandeLeadId(Integer demandeLeadId);
+    List<DetailsSpecifiques> findByDemandeLeadId(Long demandeLeadId);
 
     /**
      * Pour des statistiques : trouver toutes les réponses pour une clé précise.
