+++ src/main/java/fr/honeygroup/repository/PhotoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PhotoRepository extends JpaRepository<Photo, Integer> {
+public interface PhotoRepository extends JpaRepository<Photo, Long> {
 
 //    /**
 //     * Récupérer toutes les photos associées à une prestation.
